# contracts/managed-agent/v1alpha1/openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "Cloud Agents Managed Agent Foundation API",
    "version": "v1alpha1"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/tenants/{tenantId}/memberships": {
      "get": {
        "operationId": "managedAgentListMemberships",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/ProjectPageSize"
          },
          {
            "$ref": "#/components/parameters/ProjectPageToken"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/membership-page.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "operationId": "managedAgentCreateMembership",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../platform/v1alpha1/schemas/membership-create-request.schema.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "X-Resource-Version": {
                "$ref": "#/components/headers/ResourceVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/rbac-mutation-result.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/tenants/{tenantId}/memberships/{membershipId}": {
      "get": {
        "operationId": "managedAgentGetMembership",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/MembershipId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Resource-Version": {
                "$ref": "#/components/headers/ResourceVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/membership.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/v1/tenants/{tenantId}/role-bindings": {
      "get": {
        "operationId": "managedAgentListRoleBindings",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/ProjectPageSize"
          },
          {
            "$ref": "#/components/parameters/ProjectPageToken"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/role-binding-page.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "operationId": "managedAgentBindRole",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../platform/v1alpha1/schemas/role-binding-create-request.schema.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "X-Resource-Version": {
                "$ref": "#/components/headers/ResourceVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/rbac-mutation-result.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/tenants/{tenantId}/role-bindings/{roleBindingId}": {
      "get": {
        "operationId": "managedAgentGetRoleBinding",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/RoleBindingId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Resource-Version": {
                "$ref": "#/components/headers/ResourceVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/role-binding.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "TenantId": {
        "name": "tenantId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/identifier.schema.json"
        }
      },
...
      "ProjectPageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/page-size.schema.json"
        }
      },
      "ProjectPageToken": {
        "name": "pageToken",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/page-token.schema.json"
        }
      },
...
      "MembershipId": {
        "name": "membershipId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/identifier.schema.json"
        }
      },
...
      "RoleBindingId": {
        "name": "roleBindingId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/identifier.schema.json"
        }
      },
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/identifier.schema.json"
        }
      },
...
    },
    "headers": {
      "ResourceVersion": {
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/resource-version.schema.json"
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Problem",
        "headers": {
          "WWW-Authenticate": {
            "description": "Bearer authentication challenge returned with 401 responses.",
            "schema": {
              "$ref": "./schemas/www-authenticate.schema.json"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "../../common/v1alpha1/schemas/problem.schema.json"
            }
          }
        }
      }
    }
  },
...
}